Issue 3: Cater for UTF-8 characters in input file.
See Issue https://github.com/goodvibes2/IngAusOfxFixWin/issues/3.
The input file, as downloaded from ING Australia, *is* a UTF-8 file.
Open the input and output files with CHAR_SET UTF-8 instead of US-ASCII.
This is to fix IngAusOfxFix crashing with error message
  IOException: java.nio.charset.MalformedInputException: Input length = 1
when it finds a non-ascii UTF-8 character in the input file.
ENCODING:USASCII & CHARSET:1252 in input file are also changed to
ENCODING:UTF-8   & CHARSET:NONE respectively in the output file.
This is to allow UTF-8 characters to be imported into GnuCash.
Note that (from my limited testing) non-ascii UTF-8 characters are
imported correctly into GnuCash in Linux, but some are dropped or
incorrectly translated when importing into GnuCash in MS Windows.